Fix/editar/events

+++ src/services/events/deleteUserEvent.service.ts
@@ -3,11 +3,15 @@ import { Events } from "../../entities/event";
 import { UsersEvents } from "../../entities/userEvent";
 import { AppError } from "../../errors";
 
-const deleteUserEventService = async (eventId: string, id: string): Promise<void> => {
+const deleteUserEventService = async (
+  eventId: string,
+  id: string
+): Promise<void> => {
   const eventRepository = AppDataSource.getRepository(Events);
   const userEventRepository = AppDataSource.getRepository(UsersEvents);
 
-  if (eventId.length !== 36) throw new AppError("Id must have a valid UUID format", 400);
+  if (eventId.length !== 36)
+    throw new AppError("Id must have a valid UUID format", 400);
 
   const event: Events | null = await eventRepository.findOneBy({
     id: eventId,
@@ -15,14 +19,18 @@ const deleteUserEventService = async (eventId: string, id: string): Promise<void
 
   if (!event) throw new AppError("Event not found", 404);
 
+  const userInEvent = userEventRepository.findOneBy({ user: { id: id } });
+
+  if(!userInEvent) throw new AppError("User does not participate in this event", 404);
+
   await userEventRepository.delete({
     event: event,
     user: {
-        id:id
-    }
-  })
+      id: id,
+    },
+  });
 
-  return
+  return;
 };
 
 export default deleteUserEventService;
